Reorganize and update examples
- Well... mark some of them as broken
